hello

+++ Embedded Project/src/main.cpp
@@ -1,7 +1,7 @@
 #include <Arduino.h>
 
 // put function declarations here:
-int myFunction(int, int);
+int myFunction(float, int);
 
 void setup() {
   // put your setup code here, to run once:
@@ -12,3 +12,4 @@ void loop() {
   // put your main code here, to run repeatedly:
 }
 
+//feniusebfuoysiegfuyisekhfbsekfvc
